pyproject.toml
[project]
name = "grounding"
version = "0.1.0"
description = "MNIST grounding experiments: from naive autoencoders to robust latent spaces"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "torchaudio>=2.2",
  "numpy>=1.26",
  "matplotlib>=3.8",
  "scikit-learn>=1.4",
  "tqdm>=4.66",
  "pandas>=2.2",
  "tensorboard>=2.16",
  "jupyter>=1.0",
  "ipykernel>=6.29",
  "PyYAML>=6.0",
  "umap-learn>=0.5.6",
  "onnx>=1.20.1",
  "onnxscript>=0.6.2",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
